requirements.txt
langgraph
ollama
python-dotenv
secure-smtplib